Convert LoginResource to DRF (bug 910555)
